RAC: refactor code use new component
