+++ packages/components/src/Calendar/YearDropdown.tsx
@@ -15,20 +15,9 @@ import { Button } from '../Button';
 
 interface YearDropdownProps {
   setSelectedDropdown: Dispatch<SetStateAction<string | undefined>>;
-  setSelectedValue: Dispatch<
-    SetStateAction<{
-      month?: string;
-      year?: number;
-    }>
-  >;
-  selectedValue: { month?: string; year?: number };
 }
 
-const YearDropdown = ({
-  setSelectedDropdown,
-  setSelectedValue,
-  selectedValue,
-}: YearDropdownProps) => {
+const YearDropdown = ({ setSelectedDropdown }: YearDropdownProps) => {
   const state = useContext(CalendarStateContext)!;
   const years: { value: CalendarDate; formatted: string }[] = [];
   let formatter = useDateFormatter({
@@ -59,7 +48,6 @@ const YearDropdown = ({
     let index = Number(key);
     let date = years[index].value;
     state.setFocusedDate(date);
-    setSelectedValue({ ...selectedValue, year: date.year });
   };
 
   return (
